pyproject.toml
[tool.poetry]
name = "python-grpc-test"
version = "0.1.0"
description = ""
authors = ["nobinkhan <nazrulkhan.dk@gmail.com>"]
packages = [{include = "scripts"}]

[tool.poetry.scripts]
serveGRPC = "scripts.server:grpc_serve"
serveREST = "scripts.server:rest_serve"
client = "scripts.client:main"

[tool.poetry.dependencies]
python = "^3.12"
grpcio-tools = "^1.60.0"
fastapi = "^0.109.0"
granian = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
